--- a/build.gradle
+++ b/build.gradle
@@ -29,6 +29,7 @@ buildscript {
         classpath 'com.palantir.gradle.gitversion:gradle-git-version:0.11.0'
         classpath 'gradle.plugin.org.inferred:gradle-processors:2.2.0'
         classpath 'com.palantir.gradle.consistentversions:gradle-consistent-versions:1.9.0'
+        classpath "me.champeau.gradle:jmh-gradle-plugin:0.4.7"
     }
 }