--- a/tracing-benchmarks/build.gradle
+++ b/tracing-benchmarks/build.gradle
@@ -1,0 +1,40 @@
+/*
+ * (c) Copyright 2018 Palantir Technologies Inc. All rights reserved.
+ */
+
+//apply from: "${rootDir}/gradle/jmh.gradle"
+buildscript {
+    repositories {
+        maven {
+            url 'https://dl.bintray.com/palantir/releases/'
+        }
+        maven {
+            url 'https://plugins.gradle.org/m2/'
+        }
+    }
+    dependencies {
+        classpath "me.champeau.gradle:jmh-gradle-plugin:0.4.7"
+    }
+}
+
+apply plugin: 'me.champeau.gradle.jmh'
+
+jmh {
+    // Use profilers to collect additional data. Supported profilers:
+    // [cl, comp, gc, stack, perf, perfnorm, perfasm, xperf, xperfasm, hs_cl, hs_comp, hs_gc, hs_rt, hs_thr]
+    profilers = ['gc']
+}
+
+tasks.jmhCompileGeneratedClasses {
+    options.annotationProcessorPath = configurations.errorprone
+    options.errorprone.enabled = true
+}
+
+dependencies {
+    annotationProcessor 'org.openjdk.jmh:jmh-generator-annprocess'
+    compileOnly 'org.openjdk.jmh:jmh-generator-annprocess'
+    compile 'org.openjdk.jmh:jmh-core'
+    jmh project(':tracing')
+    jmh 'com.google.guava:guava'
+    jmh 'org.apache.logging.log4j:log4j-slf4j-impl'
+}