--- a/tracing-benchmarks/build.gradle
+++ b/tracing-benchmarks/build.gradle
@@ -32,7 +32,7 @@ tasks.jmhCompileGeneratedClasses {
 dependencies {
     annotationProcessor 'org.openjdk.jmh:jmh-generator-annprocess'
     compileOnly 'org.openjdk.jmh:jmh-generator-annprocess'
-    compile "org.immutables:value::annotations"
+    compile 'org.immutables:value::annotations'
     compile 'org.openjdk.jmh:jmh-core'
     jmh project(':tracing')
     jmh 'com.google.guava:guava'