--- a/tracing-benchmarks/build.gradle
+++ b/tracing-benchmarks/build.gradle
@@ -2,20 +2,6 @@
  * (c) Copyright 2018 Palantir Technologies Inc. All rights reserved.
  */
 
-buildscript {
-    repositories {
-        maven {
-            url 'https://dl.bintray.com/palantir/releases/'
-        }
-        maven {
-            url 'https://plugins.gradle.org/m2/'
-        }
-    }
-    dependencies {
-        classpath "me.champeau.gradle:jmh-gradle-plugin:0.4.7"
-    }
-}
-
 apply plugin: 'me.champeau.gradle.jmh'
 
 jmh {