… settings.gradle
… versions.lock
… versions.props