@wenhoujx 2019-06-18
```
Benchmark                                                              Mode  Cnt       Score       Error   Units
TracingBenchmark.raw                                                   avgt    5       8.000 ±     1.116   ns/op
TracingBenchmark.raw:·gc.alloc.rate                                    avgt    5       0.007 ±     0.018  MB/sec
TracingBenchmark.raw:·gc.alloc.rate.norm                               avgt    5      ≈ 10⁻⁵                B/op
TracingBenchmark.raw:·gc.count                                         avgt    5         ≈ 0              counts
TracingBenchmark.sampledCallable                                       avgt    5    2415.347 ±   343.585   ns/op
TracingBenchmark.sampledCallable:·gc.alloc.rate                        avgt    5    6515.868 ±   952.898  MB/sec
TracingBenchmark.sampledCallable:·gc.alloc.rate.norm                   avgt    5    1520.002 ±     0.004    B/op
TracingBenchmark.sampledCallable:·gc.churn.PS_Eden_Space               avgt    5    6657.981 ±  1158.048  MB/sec
TracingBenchmark.sampledCallable:·gc.churn.PS_Eden_Space.norm          avgt    5    1554.165 ±   282.045    B/op
TracingBenchmark.sampledCallable:·gc.churn.PS_Survivor_Space           avgt    5       0.207 ±     0.415  MB/sec
TracingBenchmark.sampledCallable:·gc.churn.PS_Survivor_Space.norm      avgt    5       0.048 ±     0.095    B/op
TracingBenchmark.sampledCallable:·gc.count                             avgt    5      62.000              counts
TracingBenchmark.sampledCallable:·gc.time                              avgt    5      49.000                  ms
TracingBenchmark.undecidedCallable                                     avgt    5    2064.025 ±   496.632   ns/op
TracingBenchmark.undecidedCallable:·gc.alloc.rate                      avgt    5    6414.531 ±  1457.700  MB/sec
TracingBenchmark.undecidedCallable:·gc.alloc.rate.norm                 avgt    5    1275.209 ±     0.080    B/op
TracingBenchmark.undecidedCallable:·gc.churn.PS_Eden_Space             avgt    5    6466.522 ±  2177.014  MB/sec
TracingBenchmark.undecidedCallable:·gc.churn.PS_Eden_Space.norm        avgt    5    1283.894 ±   176.306    B/op
TracingBenchmark.undecidedCallable:·gc.churn.PS_Survivor_Space         avgt    5       0.250 ±     0.119  MB/sec
TracingBenchmark.undecidedCallable:·gc.churn.PS_Survivor_Space.norm    avgt    5       0.050 ±     0.018    B/op
TracingBenchmark.undecidedCallable:·gc.count                           avgt    5      63.000              counts
TracingBenchmark.undecidedCallable:·gc.time                            avgt    5      50.000                  ms
TracingBenchmark.unsampledCallable                                     avgt    5    2469.110 ±   552.636   ns/op
TracingBenchmark.unsampledCallable:·gc.alloc.rate                      avgt    5    5264.609 ±  1174.790  MB/sec
TracingBenchmark.unsampledCallable:·gc.alloc.rate.norm                 avgt    5    1248.013 ±     0.097    B/op
TracingBenchmark.unsampledCallable:·gc.churn.PS_Eden_Space             avgt    5    5355.468 ±   847.513  MB/sec
TracingBenchmark.unsampledCallable:·gc.churn.PS_Eden_Space.norm        avgt    5    1272.001 ±   273.956    B/op
TracingBenchmark.unsampledCallable:·gc.churn.PS_Survivor_Space         avgt    5       0.136 ±     0.215  MB/sec
TracingBenchmark.unsampledCallable:·gc.churn.PS_Survivor_Space.norm    avgt    5       0.032 ±     0.054    B/op
TracingBenchmark.unsampledCallable:·gc.count                           avgt    5      45.000              counts
TracingBenchmark.unsampledCallable:·gc.time                            avgt    5      45.000                  ms
TracingBenchmark.sampledTracerSpan                                     avgt    5  154163.541 ± 60469.233   ns/op
TracingBenchmark.sampledTracerSpan:·gc.alloc.rate                      avgt    5    7292.102 ±  2372.250  MB/sec
TracingBenchmark.sampledTracerSpan:·gc.alloc.rate.norm                 avgt    5  108564.545 ±    13.229    B/op
TracingBenchmark.sampledTracerSpan:·gc.churn.PS_Eden_Space             avgt    5    7433.542 ±  2531.196  MB/sec
TracingBenchmark.sampledTracerSpan:·gc.churn.PS_Eden_Space.norm        avgt    5  110667.398 ± 10206.453    B/op
TracingBenchmark.sampledTracerSpan:·gc.churn.PS_Survivor_Space         avgt    5       0.230 ±     0.175  MB/sec
TracingBenchmark.sampledTracerSpan:·gc.churn.PS_Survivor_Space.norm    avgt    5       3.490 ±     3.741    B/op
TracingBenchmark.sampledTracerSpan:·gc.count                           avgt    5      64.000              counts
TracingBenchmark.sampledTracerSpan:·gc.time                            avgt    5      50.000                  ms
TracingBenchmark.unsampledTracerSpan                                   avgt    5  232746.583 ± 61347.557   ns/op
TracingBenchmark.unsampledTracerSpan:·gc.alloc.rate                    avgt    5    4880.881 ±  1344.496  MB/sec
TracingBenchmark.unsampledTracerSpan:·gc.alloc.rate.norm               avgt    5  108926.232 ±     9.732    B/op
TracingBenchmark.unsampledTracerSpan:·gc.churn.PS_Eden_Space           avgt    5    5037.797 ±  2114.684  MB/sec
TracingBenchmark.unsampledTracerSpan:·gc.churn.PS_Eden_Space.norm      avgt    5  112203.221 ± 18934.702    B/op
TracingBenchmark.unsampledTracerSpan:·gc.churn.PS_Survivor_Space       avgt    5       0.080 ±     0.108  MB/sec
TracingBenchmark.unsampledTracerSpan:·gc.churn.PS_Survivor_Space.norm  avgt    5       1.809 ±     2.714    B/op
TracingBenchmark.unsampledTracerSpan:·gc.count                         avgt    5      39.000              counts
TracingBenchmark.unsampledTracerSpan:·gc.time                          avgt    5      45.000                  ms
TracingBenchmark.undecidedTracerSpan                                   avgt    5  172844.869 ± 47191.122   ns/op
TracingBenchmark.undecidedTracerSpan:·gc.alloc.rate                    avgt    5    6543.852 ±  1775.645  MB/sec
TracingBenchmark.undecidedTracerSpan:·gc.alloc.rate.norm               avgt    5  108321.556 ±    11.003    B/op
TracingBenchmark.undecidedTracerSpan:·gc.churn.PS_Eden_Space           avgt    5    6670.054 ±  2739.824  MB/sec
TracingBenchmark.undecidedTracerSpan:·gc.churn.PS_Eden_Space.norm      avgt    5  110249.676 ± 20044.784    B/op
TracingBenchmark.undecidedTracerSpan:·gc.churn.PS_Survivor_Space       avgt    5       0.145 ±     0.086  MB/sec
TracingBenchmark.undecidedTracerSpan:·gc.churn.PS_Survivor_Space.norm  avgt    5       2.396 ±     0.759    B/op
TracingBenchmark.undecidedTracerSpan:·gc.count                         avgt    5      56.000              counts
TracingBenchmark.undecidedTracerSpan:·gc.time                          avgt    5      51.000                  ms

```